Add files via upload
This project focuses on the extraction of MFCC from audio signals in C and the calculation of Euclidean Distances between vowels to see how similar or different each vowel is from one another. While one vowel will have the lowest distance with itself, the distance will be higher in case of other vowels. Vowels with acoustic similarities will be closer compared to other vowels. The extraction of MFCCs involves several methodical steps to transform raw audio signals into a set of coefficients that effectively represent the perceptual properties of sound. The various steps of pre-emphasis, framing, FFT, Mel Filter Bank, DCT and finally MFCC extraction is shown in a simple manner in the code. In order to generate the audio file, download Cool Edit 2000, record a vowel, trim the parts of it with noise and store it in ASCII file format. Two examples of such a file is uploaded, namely "234101011_a_24-1.txt" and "234101011_e_8-1.txt". Development and execution of the code is done in Code::Blocks 20.03.